+++ src/index.js
@@ -9,7 +9,5 @@ export {default as scaleQuantize} from "./quantize";
 export {default as scaleThreshold} from "./threshold";
 export {default as scaleTime} from "./time";
 export {default as scaleUtc} from "./utcTime";
-export {default as scaleSequential} from "./sequential";
-export {default as scaleSequentialLog} from "./sequentialLog";
-export {default as scaleSequentialPow, sequentialSqrt as scaleSequentialSqrt} from "./sequentialPow";
-export {default as scaleDiverging} from "./diverging";
+export {default as scaleSequential, sequentialLog as scaleSequentialLog, sequentialPow as scaleSequentialPow, sequentialSqrt as scaleSequentialSqrt} from "./sequential";
+export {default as scaleDiverging, divergingLog as scaleDivergingLog, divergingPow as scaleDivergingPow, divergingSqrt as scaleDivergingSqrt} from "./diverging";
